Create capstone project 1
Problem Statement:

megaCart is an online shopping application, where users can buy various products based on their liking. In this application, the user needs to login to execute the user stories.

Following are the user stories : 

login : Consists of a simple form, which can be used to validate username and password and populate appropriate error message on the screen. 

sign-up : Consists of a simple form, which can be used to register them and populate appropriate error message on the screen. 

main : Would display categories of items available in the online shopping application, and the user can select the item which can be added into the cart. 

buy-item : Would display the list of items added in the cart by retrieving the associative array from the localStorage.